inotagent/pyproject.toml
[project]
name = "inotagent"
version = "0.1.0"
requires-python = ">=3.12,<3.13"
description = "Async Python agent runtime for openvaia"

dependencies = [
    "anthropic>=0.40",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "tiktoken>=0.8",
    "discord.py>=2.4",
    "slack-bolt>=1.20",
    "python-telegram-bot>=21.0",
    "psycopg[binary,pool]>=3.2",
    "playwright>=1.40",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
